requirement.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
idna==2.10
Pigments==1.6
Pillow==8.1.0
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
